v0.2.2 upload
harvestable plants have been added although the only one added was mung beans, cattails and bamboo would be added in the future.
you can now build a sticks pile by building a crossed sticks and dropping a stick to it. You can also drop a burning wood dust to the sticks pile to burn it, adding a stick turns it into a fireplace.
drop a ceramic cup into the water, you'll get a ceramic cup with saltwater, you can boil it to get drinkable water.