v0.2.3 upload
this is a very significant update to the game,

you can now mine iron ores and bamboo.
there are now animals around the map.
a craftable basket is added and stores 3 items at a time

animals have AI that help them survive. There are still only earthworms and pigeons, UI and performance would be the main focus in the next release.